refactor(net): extract reusable UPnP IGD server (igd_server.rs)
Move the SSDP/SOAP codec and the three control actions (GetExternalIPAddress,
AddPortMapping/AddAnyPortMapping, DeletePortMapping) into net/igd_server.rs,
generic over GatewayBackend, alongside the PCP server. tunnel/igd.rs keeps the
wg-bound SSDP+HTTP transport and its forward helpers, and its control handler
now delegates to igd_server::handle_control. add_forward now returns the UPnP
error code so the IGD reports 718/501 precisely (PCP maps any error to
NO_RESOURCES). igd_xml/ moved to net/. Behavior-preserving; 21 tests pass.

Completes the backend-generic gateway server (PCP + IGD) that StartWRT will
reuse with a UCI-firewall forward backend.
