+++ core/src/net/igd_xml/scpd.xml
