+++ core/src/net/mod.rs
@@ -7,6 +7,7 @@ pub mod forward;
 pub mod gateway;
 pub mod host;
 pub mod http;
+pub mod igd_server;
 pub mod keys;
 pub mod mdns;
 pub mod net_controller;
